+++ app/src/main/java/com/example/aquariumtracker/repository/ParameterRepository.kt
@@ -7,13 +7,13 @@ import com.example.aquariumtracker.database.model.ParameterWithMeasurements
 
 class ParameterRepository(private val parameterDAO: ParameterDAO) {
 
-    fun getParametersForAquarium(aqID: Int) = parameterDAO.getParametersForAquarium(aqID)
+    fun getParametersForAquarium(aqID: Long) = parameterDAO.getParametersForAquarium(aqID)
 
     val aq0Params: LiveData<List<Parameter>> = parameterDAO.getParametersForAquarium(0)
 
     val allParams: LiveData<List<Parameter>> = parameterDAO.getAllParametersList()
 
-    fun getParameterWithMeasurements(aqID: Int): LiveData<List<ParameterWithMeasurements>>
+    fun getParameterWithMeasurements(aqID: Long): LiveData<List<ParameterWithMeasurements>>
             = parameterDAO.getParameterWithMeasurements(aqID)
 
     suspend fun insert(param: Parameter) {
